Extract BoundsChecking library to separate files
The code is moved to the new location with only very few necessary
changes (removing 'static' from some functions).

It will be improved with further commits.
