+++ llvm/utils/gn/secondary/clang/lib/StaticAnalyzer/Checkers/BUILD.gn
@@ -22,6 +22,7 @@ static_library("Checkers") {
     "BitwiseShiftChecker.cpp",
     "BlockInCriticalSectionChecker.cpp",
     "BoolAssignmentChecker.cpp",
+    "BoundsChecking.cpp",
     "BuiltinFunctionChecker.cpp",
     "CStringChecker.cpp",
     "CStringSyntaxChecker.cpp",
